requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-strip-extras --output-file=requirements.txt requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via
    #   django
    #   django-allauth
bandit==1.8.6
    # via -r requirements.in
billiard==4.2.2
    # via celery
bleach==6.2.0
    # via -r requirements.in
boolean-py==5.0
    # via license-expression
cachecontrol[filecache]==0.14.3
    # via
    #   cachecontrol
    #   pip-audit
celery==5.5.3
    # via django-celery-beat
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coverage==7.11.0
    # via -r requirements.in
cron-descriptor==2.0.6
    # via django-celery-beat
cryptography==46.0.3
    # via -r requirements.in
cssselect2==0.8.0
    # via svglib
cyclonedx-python-lib==9.1.0
    # via pip-audit
defusedxml==0.7.1
    # via py-serializable
dj-database-url==3.0.1
    # via -r requirements.in
django==5.2.8
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-appconf
    #   django-celery-beat
    #   django-extensions
    #   django-filer
    #   django-fsm-2
    #   django-fsm-2-admin
    #   django-parler
    #   django-polymorphic
    #   django-reversion
    #   django-rosetta
    #   django-taggit
    #   django-timezone-field
    #   django-tinymce
    #   djangorestframework
    #   easy-thumbnails
django-allauth==65.12.1
    # via -r requirements.in
django-appconf==1.1.0
    # via django-imagekit
django-celery-beat==2.8.1
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filer==3.3.2
    # via -r requirements.in
django-fsm-2==4.1.0
    # via
    #   -r requirements.in
    #   django-fsm-2-admin
django-fsm-2-admin==2.0.1
    # via -r requirements.in
django-imagekit==6.0.0
    # via -r requirements.in
django-parler==2.3
    # via -r requirements.in
django-polymorphic==4.1.0
    # via django-filer
django-reversion==6.0.0
    # via -r requirements.in
django-rosetta==0.10.2
    # via -r requirements.in
django-taggit==6.1.0
    # via -r requirements.in
django-timezone-field==7.1
    # via django-celery-beat
django-tinymce==5.0.0
    # via -r requirements.in
django-widget-tweaks==1.5.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
easy-thumbnails[svg]==2.10.1
    # via django-filer
filelock==3.20.0
    # via cachecontrol
freetype-py==2.5.1
    # via rlpycairo
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
heroicons==2.13.0
    # via -r requirements.in
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
kombu==5.5.4
    # via celery
license-expression==30.4.4
    # via cyclonedx-python-lib
lxml==6.0.2
    # via svglib
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via cachecontrol
packageurl-python==0.17.5
    # via cyclonedx-python-lib
packaging==25.0
    # via
    #   gunicorn
    #   kombu
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
pilkit==3.0
    # via django-imagekit
pillow==12.0.0
    # via
    #   easy-thumbnails
    #   pilkit
    #   reportlab
pip-api==0.0.34
    # via pip-audit
pip-audit==2.9.0
    # via -r requirements.in
pip-requirements-parser==32.0.1
    # via pip-audit
platformdirs==4.5.0
    # via pip-audit
pluggy==1.6.0
    # via pytest
polib==1.2.0
    # via django-rosetta
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
py-serializable==2.1.0
    # via cyclonedx-python-lib
pycairo==1.28.0
    # via rlpycairo
pycparser==2.23
    # via cffi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via -r requirements.in
pyparsing==3.2.5
    # via pip-requirements-parser
pytest==8.4.2
    # via pytest-django
pytest-django==4.11.1
    # via -r requirements.in
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.3
    # via bandit
redis==6.4.0
    # via -r requirements.in
reportlab==4.4.4
    # via
    #   easy-thumbnails
    #   svglib
requests==2.32.5
    # via
    #   cachecontrol
    #   django-rosetta
    #   pip-audit
rich==14.2.0
    # via
    #   bandit
    #   pip-audit
rlpycairo==0.4.0
    # via svglib
rules==3.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
sqlparse==0.5.3
    # via django
stevedore==5.5.0
    # via bandit
svglib==1.6.0
    # via easy-thumbnails
tinycss2==1.4.0
    # via
    #   cssselect2
    #   svglib
toml==0.10.2
    # via pip-audit
typing-extensions==4.15.0
    # via cron-descriptor
tzdata==2025.2
    # via
    #   django-celery-beat
    #   kombu
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   cssselect2
    #   tinycss2
whitenoise==6.11.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
